// src/api/generated/models/StockAdjustment.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Quick stock adjustment
 */
export type StockAdjustment = {
    inventory_item_id: string;
    new_quantity: (number | string);
    reason: string;
    notes?: (string | null);
};